# Remove custom binary-conversion functions